+++ src/org/encog/solve/genetic/genes/Gene.java
@@ -3,7 +3,7 @@ package org.encog.solve.genetic.genes;
 public interface Gene extends Comparable<Gene> {
 	public void copy(Gene gene);
 	public int getId();
-	public int getInnovationId();
+	public long getInnovationId();
 	public boolean isEnabled();
 	public void setEnabled(boolean e);
 }
